+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - CGSize: The logic behind `CGSizeAdjustToCGSize` has been mergely improved.
 - Function `iOS7Delta` has been removed.
 - UIView: Getters and setters for receiver's frame are now attributes and no longer functions.
+- NSObject: Remove deprecated functions `performBlock:afterDelay:completion:` and `performBlock:afterDelay`.
 
 ### Enhancements
 
@@ -27,6 +28,7 @@
 - UIView: `matchParentConstraints` now returns an array of `NSLayoutConstraints`.
 - UIView: `loadFromNib` avoid crash when invalid nib name.
 - UIScreen: Fix `is55InchScreen` function.
+- NSObject: Function `performBlock:completion:` will now execute the completion block even if the main block is `nil`.
 
 ## 1.0.0